requirements.txt
asgiref==3.10.0
certifi==2025.11.12
cloudinary==1.44.1
Django==5.2.7
django-cors-headers==4.9.0
django-extensions==4.1
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
psycopg==3.2.10
psycopg-binary==3.2.10
PyJWT==2.10.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
